clean html in modal with useFieldHtml
