+++ app/javascript/mastodon/features/account_timeline/modals/field_modal.tsx
@@ -2,7 +2,10 @@ import type { FC } from 'react';
 
 import { FormattedMessage } from 'react-intl';
 
+import { EmojiHTML } from '@/mastodon/components/emoji/html';
+
 import type { AccountField } from '../common';
+import { useFieldHtml } from '../hooks/useFieldHtml';
 
 import classes from './styles.module.css';
 
@@ -10,12 +13,23 @@ export const AccountFieldModal: FC<{
   onClose: () => void;
   field: AccountField;
 }> = ({ onClose, field }) => {
+  const handleLabelElement = useFieldHtml(field.nameHasEmojis);
+  const handleValueElement = useFieldHtml(field.valueHasEmojis);
   return (
     <div className='modal-root__modal safety-action-modal'>
       <div className='safety-action-modal__top'>
         <div className='safety-action-modal__confirmation'>
-          <p>{field.name}</p>
-          <p className={classes.fieldValue}>{field.value}</p>
+          <EmojiHTML
+            as='p'
+            htmlString={field.name_emojified}
+            onElement={handleLabelElement}
+          />
+          <EmojiHTML
+            as='p'
+            htmlString={field.value_emojified}
+            onElement={handleValueElement}
+            className={classes.fieldValue}
+          />
         </div>
       </div>
       <div className='safety-action-modal__bottom'>
